BE/requirements.txt
alembic==1.11.1
anyio==3.6.2
asyncpg==0.27.0
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.2
fastapi-async-sqlalchemy==0.3.14
fastapi-pagination==0.12.4
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.13
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.41.1
sqlalchemy2-stubs==0.0.2a34
sqlmodel @ git+https://github.com/adpmhel24/sqlmodel.git@09dc2f437e1a9445161450eb841ef967b2177c33
starlette==0.27.0
typing_extensions==4.6.1
ujson==5.7.0
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
